Fixes #366 Histogram JList on StatsFrame now has copy transferable for the string value.
